Separated proxy server into separate file
